requirements.txt
Django==5.1.7
djangorestframework==3.15.2
whitenoise==6.9.0
pillow==11.1.0
django-tinymce==4.1.0
beautifulsoup4==4.13.3
django-ckeditor-5==0.2.17
